function loop tested.

+++ driver/arrayout.cpp
@@ -7,7 +7,7 @@
 
 #include "arrayout.hpp"
 
-void printArray(long *arrPtr, long *arrPtrMax, int n) {
+void printArray(long *arrPtr, long *arrPtrMax, long n) {
     int i(0);
     for (long *ptr(arrPtr); ptr < arrPtrMax && n > ++i; ptr++) {
         std::cout << *ptr << " ";
